Rename stdcov-service to module service
