+++ cmd/service/data_test.go
@@ -1,4 +1,4 @@
-package main
+package service
 
 import (
 	"bytes"
